[animation] move "Animanion" code in own module.
